grouped pages/* components
