--- src/App.test.tsx
@@ -1,12 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import 'jsdom-worker';
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-
-import App from './App';
-
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
